Refactor packages & classes structure
