+++ src/main/java/moe/yushi/authlibinjector/transform/CallbackSupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019  Haowei Wen <yushijinhun@gmail.com> and contributors
+ * Copyright (C) 2020  Haowei Wen <yushijinhun@gmail.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -24,8 +24,6 @@ import java.lang.reflect.Modifier;
 import org.objectweb.asm.Handle;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
-
-import moe.yushi.authlibinjector.transform.TransformUnit.TransformContext;
 
 public final class CallbackSupport {
 	private CallbackSupport() {
